+++ cmd2/rich_utils.py
@@ -202,7 +202,7 @@ class Cmd2HelpFormatter(RichHelpFormatter):
         get_metavar = self._metavar_formatter(action, default_metavar)
 
         # Handle nargs specified as a range
-        nargs_range = action.get_nargs_range()  # type: ignore[attr-defined]
+        nargs_range = action.get_nargs_range()  # type: ignore[attr-defined, ty:unresolved-attribute]
         if nargs_range is not None:
             arg_str = "%s" % get_metavar(1)  # noqa: UP031
             range_str = self._build_nargs_range_str(nargs_range)
@@ -229,7 +229,7 @@ class Cmd2HelpFormatter(RichHelpFormatter):
         get_metavar = self._metavar_formatter(action, default_metavar)
 
         # Handle nargs specified as a range
-        nargs_range = action.get_nargs_range()  # type: ignore[attr-defined]
+        nargs_range = action.get_nargs_range()  # type: ignore[attr-defined, ty:unresolved-attribute]
         if nargs_range is not None:
             yield "%s" % get_metavar(1), True  # noqa: UP031
             yield self._build_nargs_range_str(nargs_range), False